requirements.txt
amqp==5.0.1
asgiref==3.2.10
billiard==3.6.3.0
celery==5.0.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
defusedxml==0.6.0
Django==3.1.1
django-allauth==0.42.0
django-rest-auth==0.9.5
djangorestframework==3.12.0
idna==2.10
kombu==5.0.2
oauthlib==3.1.0
prompt-toolkit==3.0.7
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
vine==5.0.0
wcwidth==0.2.5
